Add documentation to Github pages (#69)
* Improved the docs

* formatting

+++ docs/source/conf.py
@@ -5,65 +5,96 @@
 
 import os
 import shutil
+import sys
+
+# Make the package importable for autodoc
+sys.path.insert(0, os.path.abspath("../../src"))
 
 
 def copy_tutorials(app):
-    src = os.path.abspath("../tutorials")
-    dst = os.path.abspath("source/tutorials")
+    confdir = os.path.dirname(os.path.abspath(__file__))
+    src = os.path.join(confdir, "..", "..", "tutorials")
+    dst = os.path.join(confdir, "tutorials")
 
-    # Remove existing target directory if it exists
     if os.path.exists(dst):
         shutil.rmtree(dst)
 
-    shutil.copytree(src, dst)
+    if os.path.exists(src):
+        shutil.copytree(src, dst)
+
+
+def copy_figures(app):
+    # conf.py lives in docs/source/, figures/ is at repo root
+    confdir = os.path.dirname(os.path.abspath(__file__))
+    src = os.path.join(confdir, "..", "..", "figures")
+    dst = os.path.join(confdir, "_static", "figures")
+
+    if os.path.exists(dst):
+        shutil.rmtree(dst)
+
+    if os.path.exists(src):
+        shutil.copytree(src, dst)
 
 
 def setup(app):
     app.connect("builder-inited", copy_tutorials)
-    # app.add_stylesheet("my-styles.css")
+    app.connect("builder-inited", copy_figures)
     app.add_css_file("custom.css")
 
 
 # -- Project information -----------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "python-template"
+project = "scope-profiler"
 copyright = "2025, Max"
 author = "Max"
 
 # -- General configuration ---------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
     "nbsphinx",
     "sphinx.ext.mathjax",
     "sphinx.ext.autodoc",
-    "myst_parser",  # enable Markdown support
+    "sphinx.ext.napoleon",
+    "sphinx.ext.intersphinx",
+    "myst_parser",
 ]
 
 exclude_patterns = []
 
-# Recognize both .rst and .md
 source_suffix = {
     ".rst": "restructuredtext",
     ".md": "markdown",
 }
 
+# MyST settings
+myst_enable_extensions = [
+    "colon_fence",
+    "fieldlist",
+]
+myst_heading_anchors = 3
+
+# Autodoc settings
+autodoc_member_order = "bysource"
+autodoc_typehints = "description"
+
+# Intersphinx
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "numpy": ("https://numpy.org/doc/stable", None),
+    "h5py": ("https://docs.h5py.org/en/stable", None),
+}
+
 # -- Options for HTML output -------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-# Activate the theme.
 html_theme = "sphinx_book_theme"
-
 
 html_static_path = ["_static"]
 templates_path = ["_templates"]
 
-
-# html_sidebars = {"**": []}
-
 html_theme_options = {
+    "repository_url": "https://github.com/max-models/scope-profiler",
     "repository_branch": "devel",
+    "use_repository_button": True,
     "show_toc_level": 3,
     "secondary_sidebar_items": ["page-toc"],
     "icon_links": [
@@ -71,6 +102,12 @@ html_theme_options = {
             "name": "GitHub",
             "url": "https://github.com/max-models/scope-profiler",
             "icon": "fab fa-github",
+            "type": "fontawesome",
+        },
+        {
+            "name": "PyPI",
+            "url": "https://pypi.org/project/scope-profiler/",
+            "icon": "fas fa-box",
             "type": "fontawesome",
         },
     ],
